Update Video to use AnimationClockComposite
